etl/clean-web-extractor-pipeline/requirements.txt
# Pipeline d'extraction web immobilière - Dépendances modernes
# Core dependencies
pydantic>=2.5.0
structlog>=23.2.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0
aiohttp>=3.9.0

# Database
pymongo>=4.6.0
motor>=3.3.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Async and concurrency
asyncio-mqtt>=0.13.0
aiofiles>=23.2.0

# Configuration and environment
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Monitoring and observability
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development tools
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
